auth middleware

+++ routes/reviews.js
@@ -2,10 +2,12 @@ import express from 'express'
 import { Book, Review } from '../models/index.js'
 import saveBook from './controllers/save_book.js'
 import searchBooks from './controllers/searchbooks.js'
+import { authMiddleware } from '../middleware/auth_middleware.js'
 
 const router = new express.Router({ mergeParams: true })
 
-router.post('/', async (req, res, next) => {
+router.post('/', authMiddleware, async (req, res, next) => {
+    const userId = res.locals.user._id
 	const { bookId } = req.params
 	// const { userId } = req.locals.user
 
@@ -22,7 +24,7 @@ router.post('/', async (req, res, next) => {
 	}
 
 	try {
-		const review = await Review.create({ ...req.body, book: bookId })
+		const review = await Review.create({ ...req.body, book: bookId, user: userId })
 		await book.reviews.push(review._id)
 		await book.save()
 	} catch (e) {
@@ -59,12 +61,16 @@ router.get('/:reviewId', async (req, res, next) => {
 	}
 })
 
-router.put('/:reviewId', async (req, res, next) => {
+router.put('/:reviewId', authMiddleware, async (req, res, next) => {
+    const userId = res.locals.user._id
 	const { reviewId } = req.params
 	const { quote, content, hashtags, image } = req.body
 
 	try {
-		await Review.findByIdAndUpdate(reviewId, {
+		const review = await Review.findById(reviewId)
+        if (review == null)return next(new Error('리뷰가 존재하지 않습니다.'))
+        if(String(review.user) !== String(userId)) return next(new Error("본인이 아닙니다."))
+        await review.updateOne({
 			quote,
 			content,
 			hashtags,
@@ -77,13 +83,16 @@ router.put('/:reviewId', async (req, res, next) => {
 	}
 })
 
-router.delete('/:reviewId', async (req, res) => {
+router.delete('/:reviewId', authMiddleware, async (req, res) => {
+    const userId = res.locals.user._id
 	const { reviewId } = req.params
 
 	// By using Document instead of Query (or Model),
 	// pre deleteOne middleware can bind the document as this
 	try {
 		const review = await Review.findById(reviewId)
+        if (review == null)return next(new Error('리뷰가 존재하지 않습니다.'))
+        if(String(review.user) !== String(userId)) return next(new Error("본인이 아닙니다."))
 		await review.deleteOne()
 
 		return res.sendStatus(202)
